requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
pyjwt[crypto]>=2.8.0
stripe>=7.0.0
